# docs/swagger.yaml
basePath: /
definitions:
  models.ChallengeGuess:
    properties:
      actualPrice:
        type: number
      carId:
        type: string
      difference:
        type: number
      guessedPrice:
        type: number
      percentage:
        type: number
      points:
        type: integer
    type: object
  models.ChallengeGuessRequest:
    properties:
      guessedPrice:
        minimum: 0
        type: number
    required:
    - guessedPrice
    type: object
  models.ChallengeResponse:
    properties:
      actualPrice:
        type: number
      carId:
        type: string
      difference:
        type: number
      guessedPrice:
        type: number
      isLastCar:
        type: boolean
      message:
        type: string
      nextCarNumber:
        type: integer
      originalUrl:
        type: string
      percentage:
        type: number
      points:
        type: integer
      sessionComplete:
        type: boolean
      totalScore:
        type: integer
    type: object
  models.ChallengeSession:
    properties:
      cars:
        items:
          $ref: '#/definitions/models.EnhancedCar'
        type: array
      completedTime:
        type: string
      currentCar:
        type: integer
      guesses:
        items:
          $ref: '#/definitions/models.ChallengeGuess'
        type: array
      isComplete:
        type: boolean
      sessionId:
        type: string
      startTime:
        type: string
      totalScore:
        type: integer
    type: object
  models.EnhancedCar:
    properties:
      auctionDetails:
        description: Flag indicating this is auction data
        type: boolean
      bodyColour:
        type: string
      bodyType:
        type: string
      description:
        description: Combined description
        type: string
      doors:
        type: string
      emissionClass:
        type: string
      engine:
        type: string
      exteriorColor:
        description: More detailed color
        type: string
      fuelType:
        type: string
      fullTitle:
        description: Complete original title for Easy mode
        type: string
      gearbox:
        type: string
      id:
        description: Standard Car fields
        type: string
      images:
        items:
          type: string
        type: array
      interiorColor:
        description: Interior details
        type: string
      keyFacts:
        description: Auction key facts
        items:
          type: string
        type: array
      location:
        description: Vehicle location e.g. "Bournemouth, Dorset, United Kingdom"
        type: string
      make:
        type: string
      mileage:
        description: Standard overview fields
        type: integer
      mileageFormatted:
        description: Enhanced Bonhams fields
        type: string
      model:
        type: string
      originalUrl:
        type: string
      owners:
        type: string
      price:
        description: Will be 0 for guessing
        type: number
      registration:
        type: string
      saleDate:
        description: Sale date in format "29 Jul 2025"
        type: string
      seats:
        type: string
      steering:
        description: '"Right-hand drive"'
        type: string
      trim:
        description: Enhanced Lookers fields
        type: string
      year:
        type: integer
    type: object
  models.GuessRequest:
    properties:
      difficulty:
        description: Default to hard for backward compatibility
        enum:
        - easy
        - hard
        type: string
      gameMode:
        enum:
        - zero
        - streak
        - challenge
        type: string
      guessedPrice:
        minimum: 0
        type: number
      listingId:
        type: string
    required:
    - gameMode
    - guessedPrice
    - listingId
    type: object
  models.GuessResponse:
    properties:
      actualPrice:
        type: number
      correct:
        type: boolean
      difference:
        type: number
      gameOver:
        type: boolean
      guessedPrice:
        type: number
      message:
        type: string
      originalUrl:
        type: string
      percentage:
        type: number
      score:
        type: integer
    type: object
  models.LeaderboardEntry:
    properties:
      date:
        type: string
      difficulty:
        description: '"easy" or "hard", defaults to "hard" for backward compatibility'
        type: string
      gameMode:
        type: string
      id:
        type: string
      name:
        maxLength: 20
        minLength: 1
        type: string
      score:
        type: integer
    required:
    - name
    type: object
  models.LeaderboardSubmissionRequest:
    properties:
      difficulty:
        description: Default to hard for backward compatibility
        enum:
        - easy
        - hard
        type: string
      gameMode:
        enum:
        - streak
        - challenge
        type: string
      name:
        maxLength: 20
        minLength: 1
        type: string
      score:
        minimum: 0
        type: integer
      sessionId:
        type: string
    required:
    - gameMode
    - name
    - score
    type: object
host: localhost:8080
info:
  contact:
    name: CarGuessr Support
    url: https://github.com/Adz-ai/carguessr/issues
  description: A fun car price guessing game with multiple game modes using real car
    data from Bonhams Auction House (Hard mode) and Lookers dealership (Easy mode).
    Features Challenge Mode with GeoGuessr-style scoring, Streak Mode, and endless
    Stay at Zero mode. Includes comprehensive security, rate limiting, leaderboards,
    mobile auto-scroll, and session-based car history tracking to prevent repetition.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://carguessr.uk
  title: CarGuessr API
  version: "2.2"
paths:
  /api/admin/cache-status:
    get:
      description: Returns information about the current cache status, age, and listing
        counts. Requires admin authentication.
      produces:
      - application/json
      responses:
        "200":
          description: cache status information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized - Admin key required'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKey: []
      summary: Get cache status information (Admin Only)
      tags:
      - admin
  /api/admin/leaderboard-status:
    get:
      description: Returns information about the leaderboard file, entry counts, and
        storage details. Requires admin authentication.
      produces:
      - application/json
      responses:
        "200":
          description: leaderboard status information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized - Admin key required'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKey: []
      summary: Get leaderboard status information (Admin Only)
      tags:
      - admin
  /api/admin/listings:
    get:
      description: Returns all car listings currently loaded in the system with full
        details including prices. This reveals all answers and is restricted to admin
        access.
      produces:
      - application/json
      responses:
        "200":
          description: count and cars array
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized - Admin key required'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKey: []
      summary: Get all available car listings (Admin Only)
      tags:
      - admin
  /api/admin/refresh-listings:
    post:
      description: Triggers a non-blocking background refresh of car listings. Supports
        mode query param (bonhams/lookers/both). Requires admin authentication and
        has a 30-minute cooldown between requests. Game continues normally during
        refresh.
      parameters:
      - default: both
        description: Refresh mode (bonhams, lookers, both)
        enum:
        - bonhams
        - lookers
        - both
        in: query
        name: mode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: refresh started, status: refreshing, note: game continues
            normally'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized - Admin key required'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited or refresh cooldown
            active'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKey: []
      summary: Manually refresh car listings (Admin Only)
      tags:
      - admin
  /api/admin/test-scraper:
    get:
      description: Tests the Bonhams scraper and returns up to 10 cars with full details.
        This is an expensive operation. Requires admin authentication.
      produces:
      - application/json
      responses:
        "200":
          description: message, count, and cars array
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized - Admin key required'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error and message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKey: []
      summary: Test the car scraper directly (Admin Only)
      tags:
      - admin
  /api/challenge/{sessionId}:
    get:
      description: Returns the current state of a challenge session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChallengeSession'
        "404":
          description: 'error: Session not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current challenge session
      tags:
      - challenge
  /api/challenge/{sessionId}/guess:
    post:
      consumes:
      - application/json
      description: Submit a price guess for the current car in challenge mode. Returns
        points based on accuracy (max 5000 points). Rate limited to 60 requests per
        minute per IP.
      parameters:
      - description: Challenge Session ID (16 alphanumeric characters)
        in: path
        name: sessionId
        required: true
        type: string
      - description: 'Price guess (max price: £10,000,000)'
        in: body
        name: guess
        required: true
        schema:
          $ref: '#/definitions/models.ChallengeGuessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: points earned, totalScore, isLastCar, message, originalUrl
          schema:
            $ref: '#/definitions/models.ChallengeResponse'
        "400":
          description: 'error: Invalid request, session complete, or price exceeds
            maximum'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Session not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a guess for challenge mode
      tags:
      - challenge
  /api/challenge/start:
    post:
      description: Starts a new 10-car challenge session with GeoGuessr-style scoring.
        Supports difficulty query param (easy/hard). Rate limited to 60 requests per
        minute per IP.
      parameters:
      - description: Difficulty mode (easy for Lookers, hard for Bonhams)
        enum:
        - easy
        - hard
        in: query
        name: difficulty
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'sessionId, cars array (10 cars with prices hidden), currentCar:
            0, totalScore: 0'
          schema:
            $ref: '#/definitions/models.ChallengeSession'
        "404":
          description: 'error: Not enough cars available for challenge mode'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a new Challenge Mode session
      tags:
      - challenge
  /api/check-guess:
    post:
      consumes:
      - application/json
      description: Submit a price guess and get feedback on accuracy, score, and game
        status. Rate limited to 60 requests per minute per IP.
      parameters:
      - description: 'Price guess data (max price: £10,000,000)'
        in: body
        name: guess
        required: true
        schema:
          $ref: '#/definitions/models.GuessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GuessResponse'
        "400":
          description: 'error: Invalid request or price exceeds maximum'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Listing not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a price guess for a car
      tags:
      - game
  /api/data-source:
    get:
      description: Returns information about both data sources (Bonhams and Lookers)
        including listing counts
      produces:
      - application/json
      responses:
        "200":
          description: data sources info with Easy/Hard mode listings
          schema:
            additionalProperties: true
            type: object
      summary: Get current data source information
      tags:
      - debug
  /api/leaderboard:
    get:
      description: Returns the leaderboard optionally filtered by game mode and difficulty,
        sorted by score (descending for challenge, ascending for streak)
      parameters:
      - description: Game mode filter (streak or challenge)
        in: query
        name: mode
        type: string
      - description: Difficulty filter (easy or hard)
        in: query
        name: difficulty
        type: string
      - description: 'Maximum number of entries to return (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LeaderboardEntry'
            type: array
      summary: Get the game leaderboard
      tags:
      - game
  /api/leaderboard/submit:
    post:
      consumes:
      - application/json
      description: Submit a score to the leaderboard for streak or challenge mode.
        Validates the score against the session data.
      parameters:
      - description: Score submission data
        in: body
        name: submission
        required: true
        schema:
          $ref: '#/definitions/models.LeaderboardSubmissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success message and leaderboard position
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid request or score validation failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a score to the leaderboard
      tags:
      - game
  /api/random-enhanced-listing:
    get:
      description: Returns a random car listing with full details, supports difficulty
        query param (easy/hard)
      parameters:
      - description: Difficulty mode (easy for Lookers, hard for Bonhams)
        enum:
        - easy
        - hard
        in: query
        name: difficulty
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EnhancedCar'
        "404":
          description: 'error: No listings available'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a random car listing with all characteristics based on difficulty
      tags:
      - game
  /api/random-listing:
    get:
      description: Returns a random car listing with the price hidden (set to 0) for
        the guessing game. Rate limited to 60 requests per minute per IP.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EnhancedCar'
        "404":
          description: 'error: No listings available'
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: 'error: Too Many Requests - Rate limited'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a random car listing for the game
      tags:
      - game
schemes:
- http
- https
securityDefinitions:
  AdminKey:
    description: Admin API key required for administrative endpoints. Can also be
      passed as 'admin_key' query parameter.
    in: header
    name: X-Admin-Key
    type: apiKey
swagger: "2.0"